+++ tools/QuietBackgroundProcesses/DevHome.QuietBackgroundProcesses.PerformanceRecorderEngine/procmon.h
@@ -19,7 +19,7 @@ struct ProcessPerformanceSummary
 
 extern "C" __declspec(dllexport) HRESULT StartMonitoringProcessUtilization(uint32_t periodInMs, void** context) noexcept;
 extern "C" __declspec(dllexport) HRESULT StopMonitoringProcessUtilization(void* context) noexcept;
-extern "C" __declspec(dllexport) HRESULT GetMonitoringProcessUtilization(void* context, ProcessPerformanceSummary** ppSummaries, uint32_t* summaryCount) noexcept;
+extern "C" __declspec(dllexport) HRESULT GetMonitoringProcessUtilization(void* context, ProcessPerformanceSummary** ppSummaries, size_t* summaryCount) noexcept;
 extern "C" __declspec(dllexport) HRESULT DeleteMonitoringProcessUtilization(void* context) noexcept;
 
 using unique_process_utilization_monitoring_thread = wil::unique_any<void*, decltype(&::DeleteMonitoringProcessUtilization), ::DeleteMonitoringProcessUtilization>;
